Add non-fullscreen output option
I've recently switched to a single 4k screen as my primary desktop,
and as a result I would like to be able to have the mythnimal output
window in just one quadrant of the screen instead of always fullscreen.
This adds the ability to have the video output windowed.

It also fixes a couple of bugs with the overlay placement that were
especially egregious when using windowed mode. Hopefully this will
also fix the problems with the overlays disappearing in fullscreen.